+++ changelog.md
@@ -1,6 +1,14 @@
 # changelog
 
 ## 2.0.2
+* `CHG` view `local f ---@type fun(x:number):boolean`
+    ```lua
+    ---before
+    function f(x: number)
+      -> boolean
+    ---after
+    local f: fun(x: number): boolean
+    ```
 * `FIX` [#558](https://github.com/sumneko/lua-language-server/issues/558)
 * `FIX` [#567](https://github.com/sumneko/lua-language-server/issues/567)
 * `FIX` [#568](https://github.com/sumneko/lua-language-server/issues/568)